nip and tuck of frequency tools
git-svn-id: https://svn.csail.mit.edu/drc@3772 c7283977-0100-402a-a91a-fa70b306dbfe

+++ software/tools/latency/src/frequency/frequency-app.cpp
@@ -53,7 +53,7 @@ void App::handleUtime(const lcm::ReceiveBuffer* rbuf, const std::string& chan, c
     drc::frequency_t out;
     out.utime = msg->utime;
     for (size_t i=0;i < freqs.size() ;i++){
-      out.frequency.push_back( (int8_t) round( freqs[i]) );
+      out.frequency.push_back( (int16_t) round( freqs[i]) );
     }
     out.num = out.frequency.size();
     std::vector< std::string> chans = frequency_->getChannels();
